Initial version of Docker for Seismic Unix
The baseimage from phusion to make sure we have proper subsystems if we need that
We use a trampoline script to make sure CWPROOT is set properly for every command

The Dockerfile has a very long RUN command for building the Seismic Unix codebase.
This is to avoid adding multiple layers of temporay data, since we have to install
bintools, X11 dev libs etc to properly build SU. These libs and temporary data are
not needed once we're done, so by combining the whole execution in one RUN, we save
100s of MB in the final image. Once there's support for squashing temporary layers,
this RUN should be split up into proper components.

The initial build will let you also run the graphical commands, provided that you
call docker with the correct parameters:
    docker run --rm -t -i -e DISPLAY=$DISPLAY -v /tmp/.X11-unix:/tmp/.X11-unix <image> <commands>

Example:
    docker run --rm -t -i -e DISPLAY=$DISPLAY -v /tmp/.X11-unix:/tmp/.X11-unix <image> suplane | suxgraph title="graph"